v2.0.0
Change logs (You may have to 'pip install Pillow'):
Checks and raises error if one of the Project/Product/Course/Website column is not filled
Raises error if 'Goal Type' is not set
The list of errors is raised in a new window
The errors are hyperlinked and can be clicked to lead to the task directly
If the selected date range is greater than 7 days, the range and total will be displayed instead of saying 'Week's total', 'Week #' etc.
Deprecated function '.append()' replaced with '.concat()' to remove the warning and future proofing.
Entered Employee key will not clear automatically. No need to enter key multiple times if we are clicking 'Submit' again.
Name of window changed to 'Timesheet Generator'
Vyoma logo added - hence window size changed to accommodate for that
Background colour and a few other aesthetics have been modified.
If a newer release of the script exists on program, the script prompts the user to download the new script. The current script can no more be run.